Added callbacks, isolating file IO ops
I added callbacks to each fragment so they can communicate
with each other and the hosting activity. I've started the
process of isolating the file IO methods from the user interface
code with the new FileIO class.

This code is not in a working state at the moment, I'm
just saving my work.